Plate + Pancake changes
Made the projectile a flat cylinder.
Made the plate a kinematic so it doesn't get knowcked when it collides with a pancake.
Still some weird collision stuff going on that I don't understand.
