+++ Pancake Flipper/Assembly-CSharp.csproj
@@ -51,8 +51,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-     <Compile Include="Assets\Plate2.cs" />
-     <Compile Include="Assets\Slingshot.cs" />
+     <Compile Include="Assets\_C#\Plate2.cs" />
+     <Compile Include="Assets\_C#\Slingshot.cs" />
  <Reference Include="Vuforia.UnityExtensions.Editor">
  <HintPath>C:/Program Files/Unity/Editor/Data/PlaybackEngines/VuforiaSupport/Managed/Editor/Vuforia.UnityExtensions.Editor.dll</HintPath>
  </Reference>
