--- Pancake Flipper/Assets/Plate.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: ee09e91404bd2e642a59627d118c5141
-timeCreated: 1522956645
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
