--- Pancake Flipper/Assets/Slingshot.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: d4202070c7bc13341965b099f7849c4d
-timeCreated: 1522957257
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
