+++ Pancake Flipper/Assets/_C#.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 1323b5cc5263cc449a9e6404557315a5
+folderAsset: yes
+timeCreated: 1522964072
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
