+++ Pancake Flipper/Assets/_C#/Plate2.cs
@@ -1,0 +1,40 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Plate2 : MonoBehaviour {
+	
+	[Header ("Set in INspector")]
+
+
+	public float speed = 1f;
+
+	public float leftAndRightEdge = 30f;
+
+	public float chanceToChangeDirections = 0.05f;
+
+
+	void Start () {
+		//dropping apples every second
+	}
+
+	void Update () {
+		//Basic Movement
+		Vector3 pos = transform.position;
+		pos.x += speed * Time.deltaTime;
+		transform.position = pos;
+
+		//changing direction
+		if (pos.x < -leftAndRightEdge) {
+			speed = Mathf.Abs (speed);
+		} else if (pos.x > leftAndRightEdge) {
+			speed = -Mathf.Abs (speed);
+		}
+	}
+
+	void FixedUpdate () {
+		if (Random.value < chanceToChangeDirections) {
+			speed *= -1;
+		}
+		}
+}
