+++ Pancake Flipper/Assets/_C#/Plate2.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 221e7a37c6cd50044b72cc6a0d4d48fb
+timeCreated: 1522956772
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
