+++ Pancake Flipper/Assets/_C#/Slingshot.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 35b0af222fff4884eb266abfcf15ef7a
+timeCreated: 1522957375
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
