+++ Pancake Flipper/Assets/_Prefabs/Pancake.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 82fa0e0800e50554a8b94f49ea3cce9d
+timeCreated: 1522964050
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
