+++ Pancake Flipper/Assets/_Prefabs/Projectile.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 92b1d151704b3604f96f95b8a7ba141e
+timeCreated: 1522957271
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
