+++ Pancake Flipper/Assets/_Scene_0.unity
@@ -113,6 +113,26 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
+--- !u!1 &10549486 stripped
+GameObject:
+  m_PrefabParentObject: {fileID: 1276207808208650, guid: ee09e91404bd2e642a59627d118c5141,
+    type: 2}
+  m_PrefabInternal: {fileID: 354177469}
+--- !u!54 &10549491
+Rigidbody:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 10549486}
+  serializedVersion: 2
+  m_Mass: 0.2
+  m_Drag: 0
+  m_AngularDrag: 0.05
+  m_UseGravity: 0
+  m_IsKinematic: 1
+  m_Interpolate: 0
+  m_Constraints: 72
+  m_CollisionDetection: 1
 --- !u!1 &75778052
 GameObject:
   m_ObjectHideFlags: 0
@@ -292,11 +312,12 @@ Prefab:
     - target: {fileID: 54409046571219534, guid: ee09e91404bd2e642a59627d118c5141,
         type: 2}
       propertyPath: m_IsKinematic
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents:
     - {fileID: 114431199440476754, guid: ee09e91404bd2e642a59627d118c5141, type: 2}
     - {fileID: 114195455735780252, guid: ee09e91404bd2e642a59627d118c5141, type: 2}
+    - {fileID: 54409046571219534, guid: ee09e91404bd2e642a59627d118c5141, type: 2}
   m_ParentPrefab: {fileID: 100100000, guid: ee09e91404bd2e642a59627d118c5141, type: 2}
   m_IsPrefabParent: 0
 --- !u!1 &1010727215
@@ -328,12 +349,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 35b0af222fff4884eb266abfcf15ef7a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  prefabProjectile: {fileID: 1386215962995176, guid: 92b1d151704b3604f96f95b8a7ba141e,
+  prefabPancake: {fileID: 1069801370460976, guid: 82fa0e0800e50554a8b94f49ea3cce9d,
     type: 2}
   velocityMult: 8
-  launchPoint: {fileID: 0}
+  launchPoint: {fileID: 256858724}
   launchPos: {x: 0, y: 0, z: 0}
-  projectile: {fileID: 0}
+  pancake: {fileID: 1069801370460976, guid: 82fa0e0800e50554a8b94f49ea3cce9d, type: 2}
   aimingMode: 0
 --- !u!135 &1010727217
 SphereCollider:
@@ -492,8 +513,8 @@ Camera:
   near clip plane: 0.3
   far clip plane: 1000
   field of view: 60
-  orthographic: 1
-  orthographic size: 16
+  orthographic: 0
+  orthographic size: 23.29601
   m_Depth: -1
   m_CullingMask:
     serializedVersion: 2
@@ -516,7 +537,7 @@ Transform:
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 1414311433}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 1, z: -10}
+  m_LocalPosition: {x: 21, y: 3, z: -50}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
