+++ Pancake Flipper/Pancake Flipper.userprefs
@@ -1,9 +1,8 @@
 ﻿<Properties StartupItem="Assembly-CSharp.csproj">
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug" PreferredExecutionTarget="Unity.Instance.Unity Editor" />
-  <MonoDevelop.Ide.Workbench ActiveDocument="Assets\Slingshot.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="Assets\_C#\Slingshot.cs">
     <Files>
-      <File FileName="Assets\Plate2.cs" Line="14" Column="46" />
-      <File FileName="Assets\Slingshot.cs" Line="75" Column="22" />
+      <File FileName="Assets\_C#\Slingshot.cs" Line="69" Column="1" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
